#412 merge and reorganize packages
