+++ src/main/java/org/jpeek/XslReport.java
@@ -39,6 +39,7 @@ import java.util.Map;
 import org.cactoos.collection.CollectionOf;
 import org.cactoos.io.TeeInput;
 import org.cactoos.scalar.LengthOf;
+import org.jpeek.calculus.Calculus;
 import org.xembly.Directives;
 import org.xembly.Xembler;
