+++ src/main/java/org/jpeek/calculus/package-info.java
@@ -21,33 +21,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.jpeek;
-
-import com.jcabi.xml.XML;
-import java.io.IOException;
-import java.util.Map;
 
 /**
- * Metrics calculus interface.
+ * JPeek calculus.
+ *
  * @since 0.30.9
- * @todo #412:30min We start implementing LCOM4Calculus implementation. We should continue
- *  calculating real 'class@value' and 'class/vars/var[@id=pairs]/text' values and remove
- *  @Disabled annotation on all Lcom4CalculusTest class. If we choose to calculate LCOM4 with Java
- *  only we should also remove LCOM4.xsl part that is doing the calculus and let the xsl here just
- *  to build the structure of the xml result. This part is especially the one calculating
- *  "xsl:variable name='E'" L73->L89, and the one doing the division L97 -> L99
  */
-public interface Calculus {
-
-    /**
-     * Produces {@link XML} representing metrics values.
-     * @param metric Desired metric to calculate
-     * @param params Params
-     * @param skeleton Package input
-     * @return XML document giving metrics values for classes
-     * @throws IOException If fails
-     */
-    XML node(String metric, Map<String, Object> params, XML skeleton)
-        throws IOException;
-
-}
+package org.jpeek.calculus;
