+++ src/test/java/org/jpeek/MetricsTest.java
@@ -27,6 +27,7 @@ import com.jcabi.matchers.XhtmlMatchers;
 import java.nio.file.Path;
 import org.cactoos.text.FormattedText;
 import org.cactoos.text.TextOf;
+import org.jpeek.calculus.xsl.XslCalculus;
 import org.jpeek.skeleton.Skeleton;
 import org.junit.jupiter.api.io.TempDir;
 import org.junit.jupiter.params.ParameterizedTest;
