+++ src/test/java/org/jpeek/XslReportTest.java
@@ -29,6 +29,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import org.cactoos.text.TextOf;
+import org.jpeek.calculus.xsl.XslCalculus;
 import org.jpeek.skeleton.Skeleton;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
